split detail table view
